requirements.txt
absl-py==0.7.1
astor==0.7.1
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.0.2
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
networkx==2.2
numpy==1.16.2
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.2
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
termcolor==1.1.0
Werkzeug==0.15.2
transforms3d==0.3.1
tqdm==4.36.1
tensorflow-gpu==1.7.0
tensorflow-plot==0.3.2
plyfile==0.7
